Move a bunch of custom types into types.py
