+++ src/quacc/recipes/dftb/_base.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
 LOG_FILE = "dftb.out"
 GEOM_FILE = "geo_end.gen"
