+++ src/quacc/recipes/espresso/bands.py
@@ -23,7 +23,7 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
     class BandsSchema(TypedDict, total=False):
         bands_pw: RunSchema
