+++ src/quacc/recipes/espresso/core.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 
     from quacc.runners.ase import OptParams
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
 BASE_SET_METAL = {
     "input_data": {
