+++ src/quacc/recipes/espresso/dos.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
     class DosSchema(TypedDict):
         static_job: RunSchema
