+++ src/quacc/recipes/espresso/phonons.py
@@ -28,7 +28,7 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
     class PhononDosSchema(TypedDict):
         relax_job: RunSchema
