+++ src/quacc/recipes/gulp/_base.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 logger = logging.getLogger(__name__)
 
 GEOM_FILE_PBC = "gulp.cif"
