+++ src/quacc/recipes/lj/core.py
@@ -20,7 +20,7 @@ if TYPE_CHECKING:
 
     from quacc.runners.ase import OptParams, VibKwargs
     from quacc.schemas._aliases.ase import OptSchema, RunSchema, VibThermoSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
 
 @job
