+++ src/quacc/recipes/newtonnet/core.py
@@ -29,7 +29,7 @@ if TYPE_CHECKING:
 
     from quacc.runners.ase import OptParams
     from quacc.schemas._aliases.ase import OptSchema, RunSchema, VibThermoSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
 
 @job
