+++ src/quacc/recipes/orca/core.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
 
     from quacc.runners.ase import OptParams
     from quacc.schemas._aliases.cclib import cclibASEOptSchema, cclibSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
 
 @job
