+++ src/quacc/recipes/qchem/ts.py
@@ -25,7 +25,7 @@ if TYPE_CHECKING:
 
     from quacc.runners.ase import OptParams
     from quacc.schemas._aliases.ase import OptSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
     class QuasiIRCSchema(OptSchema):
         initial_irc: OptSchema
