+++ src/quacc/recipes/vasp/_base.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 
     from quacc.runners.ase import OptParams
     from quacc.schemas._aliases.vasp import VaspASEOptSchema, VaspSchema
-    from quacc.utils.files import Filenames, SourceDirectory
+    from quacc.types import Filenames, SourceDirectory
 
 
 def run_and_summarize(
